+++ src/lib/stores/audio.ts
@@ -1,3 +1,4 @@
 import { writable } from 'svelte/store';
 
 export const audioContext = writable<AudioContext>();
+export const audioNode = writable<AudioWorkletNode>();
